[#1,#2] viewModel 구현 작업
